+++ node.go
@@ -9,8 +9,8 @@ import (
 	"io"
 
 	"github.com/tendermint/go-amino"
-	"github.com/tendermint/tendermint/crypto/tmhash"
-	cmn "github.com/tendermint/tendermint/libs/common"
+	"github.com/gracenoah/tendermint/crypto/tmhash"
+	cmn "github.com/gracenoah/tendermint/libs/common"
 )
 
 // Node represents a node in a Tree.