+++ proof_range.go
@@ -6,8 +6,8 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/tendermint/tendermint/crypto/tmhash"
-	cmn "github.com/tendermint/tendermint/libs/common"
+	"github.com/gracenoah/tendermint/crypto/tmhash"
+	cmn "github.com/gracenoah/tendermint/libs/common"
 )
 
 type RangeProof struct {